Replaced ReminderInfo and CaregiverInfo screens with modals. Fixed some additional formatting issues that were causing problems on Android.
