+++ screen/medications_screens/medication_adherence.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import { ScrollView, StyleSheet, Text } from 'react-native'
 
 function MedicationAdherence(props) {
