+++ screen/medications_screens/medication_reminders_list.js
@@ -1,27 +1,25 @@
-import React from 'react'
+import React, { useReducer } from 'react'
 import { StyleSheet, Text, TouchableOpacity, View } from 'react-native'
-import { AddButton, FontAwesome5Icon, ListView, MaterialCommunityIcon } from '../components'
+import { AddButton, FontAwesome5Icon, ListView } from '../components'
 import { reminderApi } from "../../api"
 import { valueHelper, alertHelper, currentUserHelper, patientHelper, patientMedicationHelper, reminderHelper, userCredentialsHelper } from "../../helpers"
 import { styles } from '../../assets/styles'
+import { ReminderModal } from '../reminders_screens'
 
 function MedicationReminder(props) {
-  const { navigation, medicationReminder } = props
-  const { currentUser, currentPatient } = currentUserHelper.getCurrentProps(props)
+  const { medicationReminder, onEdit } = props
 
-  return ( 
+  return (
     <TouchableOpacity
       activeOpacity={0.5}
       style={styles.listButton}
-      onPress={() => { navigation.navigate('ReminderScreen', { currentUser, currentPatient, reminder: medicationReminder }) }}>
-      <View style={{ flexDirection: "row", alignItems:'center', width:'95%'}}>
+      onPress={() => { onEdit(medicationReminder) }}>
+      <View style={{ flexDirection: "row", alignItems: 'center', width: '95%' }}>
         <FontAwesome5Icon size={27} name="clock" style={[styles.icon, inlineStyles.medIcon]} />
         <Text style={inlineStyles.text}>{reminderHelper.displayAction(medicationReminder)}</Text>
+        <Text style={inlineStyles.text}>{reminderDates(medicationReminder)}</Text>
       </View>
-      <View>
-        <FontAwesome5Icon size={30} name="angle-right" style={[styles.icon, inlineStyles.medIcon]} />
-      </View>
-    </TouchableOpacity> 
+    </TouchableOpacity>
   )
 
   function reminderDates(medicationReminder) {
@@ -29,9 +27,9 @@ function MedicationReminder(props) {
       case 'DailyReminder':
         return 'Daily'
       case 'WeeklyReminder':
-        return selectedDaysOfWeek(medicationReminder)
+        return `On ${selectedDaysOfWeek(medicationReminder)}`
       case 'MonthlyReminder':
-        return `Day of Month: ${reminderHelper.dayOfMonth(medicationReminder)}`
+        return `On ${reminderHelper.dayOfMonth(medicationReminder)} Day of Month`
       default:
         return ""
     }
@@ -77,13 +75,24 @@ function MedicationRemindersList(props) {
   const { navigation } = props
   const { currentPatient, currentUser } = currentUserHelper.getCurrentProps(props)
   const { patientMedication } = props
+  const [state, dispatch] = useReducer(updateState, initialState())
 
   return (
     <View style={styles.mainBody}>
-      <View style={{ display:'flex', justifyContent:'flex-end', textAlign:'right'}}>
+      <View style={{ display: 'flex', justifyContent: 'flex-end', textAlign: 'right' }}>
         <AddButton onPress={addReminder} />
       </View>
+      <ReminderModal
+        action={state.modalAction}
+        currentPatient={currentPatient}
+        currentUser={currentUser}
+        onClose={closeModal}
+        patientMedication={patientMedication}
+        reminder={state.reminder}
+        visible={state.modalVisible}
+      />
       <ListView
+        forceUpdate={state.forceUpdate}
         label='Medication Reminders'
         navigation={navigation}
         onLoadPage={loadPage}
@@ -95,7 +104,19 @@ function MedicationRemindersList(props) {
   )
 
   function addReminder() {
-    navigation.navigate('ReminderScreen', { currentUser, currentPatient, patientMedication: patientMedication })
+    dispatch({ type: 'ADD' })
+  }
+
+  function closeModal() {
+    dispatch({ type: 'CLOSE' })
+  }
+
+  function editReminder(reminder) {
+    dispatch({ type: 'EDIT', reminder })
+  }
+
+  function initialState() {
+    return { modalVisible: false }
   }
 
   function loadPage(number, size, onSuccess) {
@@ -108,7 +129,10 @@ function MedicationRemindersList(props) {
           userCredentials,
           patientHelper.id(currentPatient),
           { for_active: true, for_medication_label: patientMedicationHelper.medicationLabel(patientMedication), page: { number, size, total: 0 }, sort: 'recur_from,recur_to,action' },
-          onSuccess,
+          (page, pageHeaders) => {
+            dispatch({ type: 'UPDATED' })
+            onSuccess(page, pageHeaders)
+          },
           (error) => {
             alertHelper.error(error)
             return
@@ -123,9 +147,34 @@ function MedicationRemindersList(props) {
       <MedicationReminder
         key={`medication-reminder-${reminderHelper.id(medicationReminder)}-${medicationReminderIdx}`}
         medicationReminder={medicationReminder}
+        onEdit={editReminder}
         {...props}
       />
     )
+  }
+
+  function updateState(oldState, action) {
+    switch (action.type) {
+      case 'ADD':
+        return { ...oldState, modalVisible: true, modalAction: 'ADD' }
+
+      case 'CLOSE':
+        const newState = { ...oldState }
+        delete newState.modalAction
+        delete newState.reminder
+        newState.modalVisible = !newState.modalVisible
+        newState.forceUpdate = !newState.modalVisible
+        return newState
+
+      case 'EDIT':
+        return { ...oldState, modalVisible: true, modalAction: 'EDIT', reminder: action.reminder }
+
+      case 'UPDATED':
+        return { ...oldState, forceUpdate: false }
+
+      default:
+        return oldState
+    }
   }
 }
 
@@ -134,7 +183,7 @@ export { MedicationRemindersList }
 const inlineStyles = StyleSheet.create(
   {
     view: { flex: 1, flexDirection: "row", height: 50, margin: 5, backgroundColor: "#F3F6F9" },
-    text: { color: "#112B37", fontSize: 18, fontWeight: "500", marginLeft:5 },
+    text: { color: "#112B37", fontSize: 18, fontWeight: "500", marginLeft: 5 },
     item: { whiteSpace: 'wrap' },
     medIcon: { marginRight: 5 }
   }
